+++ lib/src/fullstory_flutter_method_channel.dart
@@ -115,17 +115,20 @@ class MethodChannelFullstoryFlutter extends FullstoryFlutterPlatform {
   // page() here asynchronously returns the ID of the native page object
   // FS.page() will use the ID future to synchronously create a new FSPage object
   @override
-  Future<int> page(String pageName, Map<String, Object?> pageVars) async {
+  Future<int> page(String pageName, Map<String, Object?> properties) async {
     final id = await methodChannel.invokeMethod<int>('page', {
       'pageName': pageName,
-      'pageVars': pageVars,
+      'properties': properties,
     });
     return id!;
   }
 
   @override
-  Future<void> startPage(int pageId) {
-    return methodChannel.invokeMethod('startPage', pageId);
+  Future<void> startPage(int pageId, Map<String, Object?> propertyUpdates) {
+    return methodChannel.invokeMethod('startPage', {
+      'pageId': pageId,
+      'propertyUpdates': propertyUpdates,
+    });
   }
 
   @override