+++ lib/src/fullstory_flutter_platform_interface.dart
@@ -83,11 +83,11 @@ abstract class FullstoryFlutterPlatform extends PlatformInterface {
   }
   // todo: webview injection disable
 
-  Future<int> page(String pageName, Map<String, dynamic> pageVars) {
+  Future<int> page(String pageName, Map<String, Object?> properties) {
     throw UnimplementedError();
   }
 
-  Future<void> startPage(int pageId) {
+  Future<void> startPage(int pageId, Map<String, Object?> propertyUpdates) {
     throw UnimplementedError();
   }
 
@@ -96,7 +96,7 @@ abstract class FullstoryFlutterPlatform extends PlatformInterface {
   }
 
   Future<void> updatePageProperties(
-      int pageId, Map<String, dynamic> properties) {
+      int pageId, Map<String, Object?> properties) {
     throw UnimplementedError();
   }